updated feed to be profile
